requirements.txt
numpy>=2.2.3
Pillow>=11.1.0
scikit-image>=0.25.1
scikit-learn>=1.5.1